# TvShowTracker/Model/DTO/ActorDTO.cs
namespace TvShowTracker.Model
{
    public class ActorDTO
    {
        public int Id { get; set; }
        public string? Name { get; set; }
        public string? Character { get; set; }
        public string? Picture { get; set; }
    }
}