# openApi.yaml
openapi: 3.0.1
info:
  title: TvShowTracker
  version: v1
paths:
...
  /v1/TVShows:
    get:
      tags:
        - TVShows
      parameters:
        - name: $page
          in: query
          schema:
            type: integer
            format: int32
        - name: $pagesize
          in: query
          schema:
            type: integer
            format: int32
        - name: $sortby
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Server Error
      security:
        - Bearer: []
  /v1/TVShows/{id}:
    get:
      tags:
        - TVShows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/TVShow'
            application/json:
              schema:
                $ref: '#/components/schemas/TVShow'
            text/json:
              schema:
                $ref: '#/components/schemas/TVShow'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Server Error
      security:
        - Bearer: []
  /v1/TVShows/{id}/Actors:
    get:
      tags:
        - TVShows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: $page
          in: query
          schema:
            type: integer
            format: int32
        - name: $pagesize
          in: query
          schema:
            type: integer
            format: int32
        - name: $sortby
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Server Error
      security:
        - Bearer: []
  /v1/TVShows/{id}/Episodes:
    get:
      tags:
        - TVShows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: $page
          in: query
          schema:
            type: integer
            format: int32
        - name: $pagesize
          in: query
          schema:
            type: integer
            format: int32
        - name: $sortby
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Server Error
      security:
        - Bearer: []
  /v1/TVShows/{id}/Episodes/{idEpisode}:
    get:
      tags:
        - TVShows
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
        - name: idEpisode
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Server Error
      security:
        - Bearer: []
  /v1/Users/{id}:
    get:
      tags:
        - Users
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: Success
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Server Error
      security:
        - Bearer: []
  /v1/Users:
    post:
      tags:
        - Users
      requestBody:
        content:
          application/json-patch+json:
            schema:
              $ref: '#/components/schemas/UserPostDTO'
          application/json:
            schema:
              $ref: '#/components/schemas/UserPostDTO'
          text/json:
            schema:
              $ref: '#/components/schemas/UserPostDTO'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UserPostDTO'
      responses:
        '201':
          description: Success
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          description: Client Error
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Server Error
      security:
        - Bearer: []
components:
  schemas:
    Actor:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        character:
          type: string
          nullable: true
        picture:
          type: string
          nullable: true
        episodesActor:
          type: array
          items:
            $ref: '#/components/schemas/EpisodeActor'
          nullable: true
      additionalProperties: false
    Episode:
      type: object
      properties:
        id:
          type: integer
          format: int32
        chapter:
          type: integer
          format: int32
          nullable: true
        name:
          type: string
          nullable: true
        overview:
          type: string
          nullable: true
        posterPath:
          type: string
          nullable: true
        seasonId:
          type: integer
          format: int32
        season:
          $ref: '#/components/schemas/Season'
        episodeActors:
          type: array
          items:
            $ref: '#/components/schemas/EpisodeActor'
          nullable: true
      additionalProperties: false
    EpisodeActor:
      type: object
      properties:
        id:
          type: integer
          format: int32
        episodeId:
          type: integer
          format: int32
        episode:
          $ref: '#/components/schemas/Episode'
        actorId:
          type: integer
          format: int32
        actor:
          $ref: '#/components/schemas/Actor'
      additionalProperties: false
    FavoriteTvShow:
      type: object
      properties:
        id:
          type: integer
          format: int32
        tvShowId:
          type: integer
          format: int32
        tvShow:
          $ref: '#/components/schemas/TVShow'
        userId:
          type: integer
          format: int32
        user:
          $ref: '#/components/schemas/User'
      additionalProperties: false
...
    Genre:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        tvShowGenres:
          type: array
          items:
            $ref: '#/components/schemas/TvShowGenre'
          nullable: true
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: {}
    Season:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        seasonNumber:
          type: integer
          format: int32
        episodeCount:
          type: integer
          format: int32
        overview:
          type: string
          nullable: true
        posterPath:
          type: string
          nullable: true
        tvShowId:
          type: integer
          format: int32
        tvShow:
          $ref: '#/components/schemas/TVShow'
        episodes:
          type: array
          items:
            $ref: '#/components/schemas/Episode'
          nullable: true
      additionalProperties: false
    TVShow:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        overview:
          type: string
          nullable: true
        popularity:
          type: number
          format: double
          nullable: true
        vote_average:
          type: number
          format: double
          nullable: true
        vote_count:
          type: number
          format: double
          nullable: true
        favoriteTvShows:
          type: array
          items:
            $ref: '#/components/schemas/FavoriteTvShow'
          nullable: true
        seasons:
          type: array
          items:
            $ref: '#/components/schemas/Season'
          nullable: true
        tvShowGenres:
          type: array
          items:
            $ref: '#/components/schemas/TvShowGenre'
          nullable: true
      additionalProperties: false
    TvShowGenre:
      type: object
      properties:
        id:
          type: integer
          format: int32
        tvShowId:
          type: integer
          format: int32
        tvShow:
          $ref: '#/components/schemas/TVShow'
        genreId:
          type: integer
          format: int32
        genre:
          $ref: '#/components/schemas/Genre'
      additionalProperties: false
    User:
      type: object
      properties:
        id:
          type: integer
          format: int32
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        favoritesTvShows:
          type: array
          items:
            $ref: '#/components/schemas/FavoriteTvShow'
          nullable: true
      additionalProperties: false
    UserPostDTO:
      required:
        - email
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          maxLength: 255
          minLength: 6
          type: string
          format: password
          nullable: true
      additionalProperties: false
  securitySchemes:
    Bearer:
      type: http
      description: Put **_ONLY_** your JWT Bearer token on textbox below!
      scheme: bearer
      bearerFormat: JWT